+++ src/lib.rs
@@ -13,6 +13,9 @@ pub mod wall;
 #[path = "mahjong/yaku.rs"]
 pub mod yaku;
 
+#[path = "mahjong/river.rs"]
+pub mod river;
+
 use pyo3::prelude::*;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
